pyproject.toml
[tool.poetry]
name = "snailtrace"
version = "0.1.2"
description = "Convert raster images to vectors"
authors = ["DJStomp <dj@deepai.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.2"
Pillow = "^10.1.0"
svg-path = "^6.3"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
trace = "snailtrace.snailtrace:main"
